_Conf current dir update

+++ classification/tf_configurator/_base.py
@@ -19,7 +19,10 @@ class _Funcs(object):
 
     def __init__(self,path=None, path_name=FNAME, conf_name=CNAME):
         self.name = path_name
-        self.pdir = path
+        if isinstance(path,type(None)):
+            self.pdir = self._reader(self.name)
+        else:
+            self.pdir = path
         self.conf = conf_name
 
     @staticmethod
@@ -84,7 +87,8 @@ class _Conf(_Funcs):
         elif os.path.exists(cls.FNAME):
             raise FileExistsError(f"'{cls.FNAME}' already exists. Try to reset()")
         cls._writer(path,cls.FNAME)
-        os.mkdir(path)
+        if path:
+            os.mkdir(path)
         return cls(path)
 
     @classmethod
@@ -96,10 +100,14 @@ class _Conf(_Funcs):
         try:
             c_dir = cls._reader(path)
             os.remove(path)
-            [os.remove(c_dir+f) for f in os.listdir(c_dir)]
-            os.rmdir(c_dir)
+            if c_dir:
+                [os.remove(c_dir+f) for f in os.listdir(c_dir)]
+                os.rmdir(c_dir)
+            else:
+                pass
+                # [os.remove(c_dir+f) for f in os.listdir()]
         except FileNotFoundError:
-            print('File/Directory Not Found')
+            print('Reset: File/Directory Not Found')
 
     def add_batch(self,batch):
         assert isinstance(batch,int)
@@ -122,18 +130,33 @@ class _Conf(_Funcs):
 
     def set_config_params(self):
         conf = self._load_conf()
+
+        self._conf_batch(conf)
+        self._conf_steps(conf)
+        self._conf_train(conf)
+
+        self._writer(conf,self.pdir+self.conf)
+
+    def _conf_batch(self,conf):
         if hasattr(self,'batch'):
             conf['train']['batch'] = self.batch
+
+    def _conf_steps(self,conf):
+        auto = lambda : conf['train']['size']//conf['train']['batch']
         if hasattr(self,'steps'):
             conf['meta']['steps'] = self.steps
             if self.steps:
                 conf['train']['steps'] = self.steps
             else:
                 try:
-                    conf['train']['steps'] = conf['train']['batch']//conf['train']['size']
-                    conf['meta']['steps'] = 'Auto'
+                    conf['train']['steps'] = auto()
+                    conf['meta']['steps']  = 'Auto'
                 except:
                     pass
+        elif hasattr(self,'batch') and isinstance(conf['meta']['steps'],str):
+            conf['train']['steps'] = auto()
+    
+    def _conf_train(self,conf):
         if hasattr(self,'train'):
             conf['meta']['train'] = self.train
             if not hasattr(self,'_xyname'):
@@ -144,19 +167,15 @@ class _Conf(_Funcs):
                 conf['test']['name']  = self._xyname
         elif not isinstance(conf['meta']['train'],bool):
             raise ValueError("'train' param undefined")
-        if not self.pdir:
-            self.pdir = self._reader(self.name)
-        self._writer(conf,self.pdir+self.conf)
 
     def _load_conf(self):
         """ load config """
-        if not self.pdir:
-            self.pdir = self._reader(self.name)
-        if os.path.exists(self.pdir) and not os.path.exists(self.pdir+self.conf):
+        if not os.path.exists(self.pdir+self.conf):
             self._writer(self._empty_conf(),self.pdir+self.conf)
         return self._reader(self.pdir+self.conf)
 
     def _fname(self,train,name=None):
+        """ data file name suffix """
         if name:
             self._xyname = name
         elif train:
